added uint64_t session id (not unsigned int) it is concatenation of two uint32 random nbrs

+++ client_TCP_lib.c
@@ -103,7 +103,7 @@ int TCP_client_send_DATA(int socket_fd, my_vec_t *vec, uint64_t session_id)
     return SUCCESS;
 }
 
-void TCP_client_handler(int socket_fd, struct sockaddr_in *server_address, my_vec_t *vec, unsigned int session_id)
+void TCP_client_handler(int socket_fd, struct sockaddr_in *server_address, my_vec_t *vec, uint64_t session_id)
 {
     // Connect to the server.
     if (connect(socket_fd, (struct sockaddr *) server_address,
